tests/submission/PhilipTaylor/canvas/2d.imageData.create2.nonfinite.html

<!DOCTYPE html>
<title>Canvas test: 2d.imageData.create2.nonfinite</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.imageData.html">imageData</a>.<a href="index.2d.imageData.create2.html">create2</a>.nonfinite</h1>
<p class="desc">createImageData() throws NOT_SUPPORTED_ERR if arguments are not finite</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.imageData.getcreate.nonfinite">2d.imageData.getcreate.nonfinite</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("createImageData() throws NOT_SUPPORTED_ERR if arguments are not finite");
_addTest(function(canvas, ctx) {

assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(Infinity, 10); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(-Infinity, 10); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(NaN, 10); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(10, Infinity); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(10, -Infinity); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(10, NaN); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(Infinity, Infinity); });
var posinfobj = { valueOf: function() { return Infinity; } },
    neginfobj = { valueOf: function() { return -Infinity; } },
    nanobj = { valueOf: function() { return -Infinity; } };
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(posinfobj, 10); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(neginfobj, 10); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(nanobj, 10); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(10, posinfobj); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(10, neginfobj); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(10, nanobj); });
assert_throws("NOT_SUPPORTED_ERR", function() { ctx.createImageData(posinfobj, posinfobj); });


});
</script>
